refactor(cfloat): split cfloat into core / manipulators / iostream / debug layers (#1334)
Phase 2's reference type, following the posit pattern from #1390 and the integer one
from #1407.

  header                              lines  headers   I/O
  cfloat/cfloat.hpp   (before)       114316      489     5
  cfloat/core.hpp     (new)           78330      365     0

Criterion 1 met. Criterion 2 is not, and will not be for this type: cfloat's +=, *= and
/= construct blocktriple locals, so its core is bounded by blocktriple's ~69,800 lines
exactly as posit's 77,430 is. Recorded as aspirational on #1334.

What left the core:

  to_decimal_fixpnt_string, parse, to_binary/to_native/to_triple -> manipulators.hpp
  operator<< / operator>>                                        -> iostream.hpp (new)
  constexprClassParameters(), showLimbs(), ReportCfloatClassParameters -> debug.hpp (new)
  support/decimal.hpp (80048 lines, 4 I/O)  followed to_decimal_fixpnt_string, its only
                                            user in all of cfloat
  internal/blockbinary/manipulators.hpp     text layer, unused by the arithmetic
  <regex>, <sstream>                        left with parse()
  native/ieee754.hpp -> ieee754_core.hpp    the bit-manipulation half from #1400,
                                            62505 lines and 4 I/O down to 43486 and 0
  8 std::cerr diagnostics -> fprintf(stderr, ...), the Phase 0 idiom

The six `if constexpr (_trace_*)` sites became `#if CFLOAT_TRACE_ENABLED` blocks: a
discarded `if constexpr` branch still requires std::cout to be DECLARED, which would
keep <iostream> in every translation unit. Same reason posit's became preprocessor
guards. The behavioural switches (CFLOAT_ENABLE_LITERALS, CFLOAT_THROW_ARITHMETIC_
EXCEPTION, CFLOAT_EXCEPT, CFLOAT_NATIVE_SQRT) move from the umbrella into cfloat_impl.hpp
beside the #if blocks that test them, or including core.hpp directly would silently lose
the literal operators -- the trap #1390 hit.

The string constructor stays in the core: assign() parses "0b..." with plain std::string
operations. The regex/istringstream parse(), reachable only from operator>>, is the text
layer's -- declared in cfloat_fwd.hpp, defined in manipulators.hpp, exactly as posit does.

Two #1389 gaps surfaced and are fixed here because the split exposed them:
  - native/nonconstexpr/gcc_long_double.hpp called frexp/frexpl unqualified inside
    namespace sw::universal, relying on ieee754.hpp to drag ::frexp into scope. Adds
    <cmath> and qualifies; frexpl/ldexpl become std::frexp/std::ldexp, since <cmath>
    overloads those for long double and the C names are global-only.
  - internal/blocktriple/manipulators.hpp used to_binary(blocksignificand) without
    including blocksignificand/manipulators.hpp.

Gated on compiling AND running core.hpp alone, core.hpp + manipulators.hpp, and the
trace path, on gcc and clang. GATE 2 caught a real defect during development: to_triple
delegates to to_triple(blocktriple), and the umbrella had been supplying blocktriple's
operator<< via blocktriple_debug.hpp -- 19 cfloat tests failed until the stream and
debug layers picked those up. The umbrella building proves nothing.

Verified: cfloat suite 97 pass 0 fail on gcc; 144 pass 0 fail across cfloat, posit and
integer on clang; every api.cpp plus the whole block layer 102 pass 0 fail. ASCII clean.

+++ include/sw/universal/native/nonconstexpr/gcc_long_double.hpp
@@ -1,5 +1,6 @@
 #pragma once
 #include <string>
+#include <cmath>    // std::frexp / std::ldexp, previously arriving transitively (#1389)
 #include <cstdio>
 // gcc_long_double.hpp: nonconstexpr implementation of IEEE-754 long double manipulators
 //
@@ -253,12 +254,12 @@ inline void extract_fp_components(long double fp, bool& _sign, int& _exponent, l
 	if constexpr (std::numeric_limits<long double>::digits <= 64) {
 		if constexpr (sizeof(long double) == 8) { // it is just a double
 			_sign = fp < 0.0 ? true : false;
-			_fr = frexp(double(fp), &_exponent);
+			_fr = std::frexp(double(fp), &_exponent);
 			_fraction = uint64_t(0x000FFFFFFFFFFFFFull) & reinterpret_cast<uint64_t&>(_fr);
 		}
 		else if constexpr (sizeof(long double) == 16 && std::numeric_limits<long double>::digits <= 64) {
 			_sign = fp < 0.0 ? true : false;
-			_fr = frexpl(fp, &_exponent);
+			_fr = std::frexp(fp, &_exponent);
 			_fraction = uint64_t(0x7FFFFFFFFFFFFFFFull) & reinterpret_cast<uint64_t&>(_fr); // 80bit extended format only has 63bits of fraction
 		}
 	}
@@ -272,12 +273,12 @@ inline void extract_fp_components(long double fp, bool& _sign, int& _exponent, l
 	if (std::numeric_limits<long double>::digits <= 64) {
 		if (sizeof(long double) == 8) { // it is just a double
 			_sign = fp < 0.0 ? true : false;
-			_fr = frexp(double(fp), &_exponent);
+			_fr = std::frexp(double(fp), &_exponent);
 			_fraction = uint64_t(0x000FFFFFFFFFFFFFull) & reinterpret_cast<uint64_t&>(_fr);
 		}
 		else if (sizeof(long double) == 16 && std::numeric_limits<long double>::digits <= 64) {
 			_sign = fp < 0.0 ? true : false;
-			_fr = frexpl(fp, &_exponent);
+			_fr = std::frexp(fp, &_exponent);
 			_fraction = uint64_t(0x7FFFFFFFFFFFFFFFull) & reinterpret_cast<uint64_t&>(_fr); // 80bit extended format only has 63bits of fraction
 		}
 	}
